asdfasdfasdf

+++ GLAPIBasic/DTOs/UserInfo.cs
@@ -2,7 +2,7 @@
 {
     public class UserInfo
     {
-        public uint UserId { get; set; }
+        public long UserId { get; set; }
         public string? Address { get; set; }
         public string Name { get; set; } = null!;
         public string? CompanyName { get; set; }
